requirements.txt
elasticsearch>=8.0.0
python-dotenv>=0.15.0
requests>=2.25.0
tqdm>=4.60.0